Refactor: move strategies into their own packages
